+++ Actor/Theme.php
@@ -5,23 +5,25 @@ namespace LinkCMS\Actor;
 class Theme {
     var $info;
     var $name;
-    const THEME_DIR = __DIR__ . '/../Themes/';
+    var $directory;
+    const THEME_DIR = __DIR__ . '/../Theme/';
 
     public function __construct(String $name, $exception=true) {
         $this->name = $name;
         if (file_exists(self::THEME_DIR . $name . '/theme.json')) {
             $this->info = json_decode(file_get_contents(self::THEME_DIR . $name . '/theme.json'));
+            Display::add_template_directory(self::THEME_DIR . $name);
+            $this->initialize_theme();
         } else {
             if ($exception) {
                 throw new \Exception('Theme ' . $name . ' is missing required theme.json file. Please check that your theme is correctly installed.');
             }
         }
-        if (file_exists(self::THEME_DIR . $name . '/theme.php')) {
-            require_once(self::THEME_DIR . $name . '/theme.php');
-        } else {
-            if ($exception) {
-                throw new \Exception('Theme ' . $name . ' is missing required theme.php file. Please check that your theme is correctly installed.');
-            }
+    }
+
+    private function initialize_theme() {
+        if (isset($this->info->init)) {
+            call_user_func($this->info->init);
         }
     }